+++ testData/generator/SimpleSealedTypes.PSI.expected.java
@@ -1,0 +1,150 @@
+// ---- GeneratedTypes.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated;
+
+import com.intellij.psi.tree.IElementType;
+import com.intellij.psi.PsiElement;
+import com.intellij.lang.ASTNode;
+import generated.psi.impl.*;
+
+public interface GeneratedTypes {
+
+  IElementType BAR = new IElementType("BAR", null);
+  IElementType BAZ = new IElementType("BAZ", null);
+
+
+  class Factory {
+    public static PsiElement createElement(ASTNode node) {
+      IElementType type = node.getElementType();
+      if (type == BAR) {
+        return new BarImpl(node);
+      }
+      else if (type == BAZ) {
+        return new BazImpl(node);
+      }
+      throw new AssertionError("Unknown element type: " + type);
+    }
+  }
+}
+// ---- Bar.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public non-sealed interface Bar extends Foo {
+
+}
+// ---- Baz.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public non-sealed interface Baz extends Foo {
+
+}
+// ---- Foo.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated.psi;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElement;
+
+public sealed interface Foo extends PsiElement permits Bar, Baz {
+
+}
+// ---- BarImpl.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated.psi.impl;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.lang.ASTNode;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.util.PsiTreeUtil;
+import static generated.GeneratedTypes.*;
+import com.intellij.extapi.psi.ASTWrapperPsiElement;
+import generated.psi.*;
+
+public class BarImpl extends ASTWrapperPsiElement implements Bar {
+
+  public BarImpl(@NotNull ASTNode node) {
+    super(node);
+  }
+
+  public void accept(@NotNull Visitor visitor) {
+    visitor.visitBar(this);
+  }
+
+  @Override
+  public void accept(@NotNull PsiElementVisitor visitor) {
+    if (visitor instanceof Visitor) accept((Visitor)visitor);
+    else super.accept(visitor);
+  }
+
+}
+// ---- BazImpl.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated.psi.impl;
+
+import java.util.List;
+import org.jetbrains.annotations.*;
+import com.intellij.lang.ASTNode;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.util.PsiTreeUtil;
+import static generated.GeneratedTypes.*;
+import com.intellij.extapi.psi.ASTWrapperPsiElement;
+import generated.psi.*;
+
+public class BazImpl extends ASTWrapperPsiElement implements Baz {
+
+  public BazImpl(@NotNull ASTNode node) {
+    super(node);
+  }
+
+  public void accept(@NotNull Visitor visitor) {
+    visitor.visitBaz(this);
+  }
+
+  @Override
+  public void accept(@NotNull PsiElementVisitor visitor) {
+    if (visitor instanceof Visitor) accept((Visitor)visitor);
+    else super.accept(visitor);
+  }
+
+}
+// ---- Visitor.java -----------------
+// This is a generated file. Not intended for manual editing.
+package generated.psi;
+
+import org.jetbrains.annotations.*;
+import com.intellij.psi.PsiElementVisitor;
+import com.intellij.psi.PsiElement;
+
+public class Visitor extends PsiElementVisitor {
+
+  public void visitBar(@NotNull Bar o) {
+    visitFoo(o);
+  }
+
+  public void visitBaz(@NotNull Baz o) {
+    visitFoo(o);
+  }
+
+  public void visitFoo(@NotNull Foo o) {
+    visitPsiElement(o);
+  }
+
+  public void visitPsiElement(@NotNull PsiElement o) {
+    visitElement(o);
+  }
+
+}